+++ test/constraint_validations/tag_extensions_test.rb
@@ -21,7 +21,7 @@ class ConstraintValidations::AriaTagExtensionsTest < ActiveSupport::TestCase
     attribute :published, :boolean
 
     validates :content, presence: true, length: {maximum: 100}
-    validates :published, exclusion: {in: [true]}
+    validates :published, presence: true, exclusion: {in: [true]}
   end
 
   test "#render encodes validation attributes onto the element" do
@@ -45,7 +45,7 @@ class ConstraintValidations::AriaTagExtensionsTest < ActiveSupport::TestCase
     assert_select "input[type=?][data-validation-messages]", "hidden", count: 0
   end
 
-  test "#render encodes Active Model validation message translations" do
+  test "#render encodes Active Model validation message translations into text fields" do
     render locals: {message: Message.new}, inline: <<~ERB
       <%= form_with model: message, url: "#" do |form| %>
         <%= form.text_field :content %>
@@ -54,6 +54,21 @@ class ConstraintValidations::AriaTagExtensionsTest < ActiveSupport::TestCase
 
     assert_select "input", count: 1 do |input, *|
       messages = JSON.parse(input["data-validation-messages"])
+
+      assert messages["badInput"] = "is invalid"
+      assert messages["valueMissing"] = "can't be blank"
+    end
+  end
+
+  test "#render encodes Active Model validation message translations into select elements" do
+    render locals: {message: Message.new}, inline: <<~ERB
+      <%= fields model: message do |form| %>
+        <%= form.select :content, [true, false] %>
+      <% end %>
+    ERB
+
+    assert_select "select", count: 1 do |select, *|
+      messages = JSON.parse(select["data-validation-messages"])
 
       assert messages["badInput"] = "is invalid"
       assert messages["valueMissing"] = "can't be blank"
@@ -92,7 +107,7 @@ class ConstraintValidations::AriaTagExtensionsTest < ActiveSupport::TestCase
     assert_select "input[type=?][aria-describedby]", "hidden", count: 0
   end
 
-  test "#render encodes aria-describedby reference to the validation message element when the field is invalid" do
+  test "#render encodes aria-describedby reference to the validation message element when the text field is invalid" do
     message = Message.new.tap(&:validate)
 
     render locals: {message: message}, inline: <<~ERB
@@ -104,6 +119,20 @@ class ConstraintValidations::AriaTagExtensionsTest < ActiveSupport::TestCase
 
     assert_select "input[aria-describedby~=?]", "#{@object_name}_content_validation_message", count: 1
     assert_select "span[id=?]", "#{@object_name}_content_validation_message", count: 1
+  end
+
+  test "#render encodes aria-describedby reference to the validation message element when the select field is invalid" do
+    message = Message.new.tap(&:validate)
+
+    render locals: {message: message}, inline: <<~ERB
+      <%= form_with model: message, url: "#" do |form| %>
+        <%= form.select :published, [true, false], prompt: true %>
+        <%= form.validation_message :published %>
+      <% end %>
+    ERB
+
+    assert_select "select[aria-describedby~=?]", "#{@object_name}_published_validation_message", count: 1
+    assert_select "span[id=?]", "#{@object_name}_published_validation_message", count: 1
   end
 
   test "#render prepends to existing aria-describedby values when the field is invalid" do
@@ -121,7 +150,7 @@ class ConstraintValidations::AriaTagExtensionsTest < ActiveSupport::TestCase
     assert_select "input[aria-describedby~=?]", "other_description", count: 1
   end
 
-  test "#render encodes aria-errormessage reference with form builder namespace" do
+  test "#render encodes aria-errormessage reference into text field with form builder namespace" do
     render locals: {message: Message.new}, inline: <<~ERB
       <%= fields model: message, namespace: "namespace" do |form| %>
       <%= form.text_field :content %>
@@ -131,6 +160,18 @@ class ConstraintValidations::AriaTagExtensionsTest < ActiveSupport::TestCase
 
     assert_select "input[aria-errormessage~=?]", "namespace_#{@object_name}_content_validation_message", count: 1
     assert_select "span[id=?]", "namespace_#{@object_name}_content_validation_message", count: 1
+  end
+
+  test "#render encodes aria-errormessage reference into select with form builder namespace" do
+    render locals: {message: Message.new}, inline: <<~ERB
+      <%= fields model: message, namespace: "namespace" do |form| %>
+      <%= form.select :published, [true, false], prompt: true %>
+      <%= form.validation_message :published %>
+      <% end %>
+    ERB
+
+    assert_select "select[aria-errormessage~=?]", "namespace_#{@object_name}_published_validation_message", count: 1
+    assert_select "span[id=?]", "namespace_#{@object_name}_published_validation_message", count: 1
   end
 
   test "#render omits aria-errormessage from [type=hidden]" do
@@ -196,7 +237,7 @@ class ConstraintValidations::AriaTagExtensionsTest < ActiveSupport::TestCase
     assert_select "span[id=?]", "namespace_#{@object_name}_nested_1_content_validation_message", count: 1
   end
 
-  test "#render sets aria-invalid when the instance is invalid" do
+  test "#render sets aria-invalid on the text field when the instance is invalid" do
     message = Message.new.tap(&:validate)
 
     render locals: {message: message}, inline: <<~ERB
@@ -206,6 +247,31 @@ class ConstraintValidations::AriaTagExtensionsTest < ActiveSupport::TestCase
     ERB
 
     assert_select "input[type=?][aria-invalid=?]", "text", "true", count: 1
+  end
+
+  test "#render sets aria-invalid on the select when the instance is invalid" do
+    message = Message.new.tap(&:validate)
+
+    render locals: {message: message}, inline: <<~ERB
+      <%= form_with model: message, url: "#" do |form| %>
+        <%= form.select :published, [true, false], prompt: true %>
+      <% end %>
+    ERB
+
+    assert_select "select[required][aria-invalid=?]", "true", count: 1
+  end
+
+  test "#render does not override [required] on the select" do
+    message = Message.new
+
+    render locals: {message: message}, inline: <<~ERB
+      <%= fields model: message do |form| %>
+        <%= form.select :published, [true, false], {prompt: true}, required: false %>
+      <% end %>
+    ERB
+
+    assert_select "select:not([required])", count: 1
+    assert_select "select[required]", count: 0
   end
 
   test "#render omits aria-invalid when the [type=hidden]" do